backend/requirements.txt
# FastAPI Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.28
psycopg2-binary>=2.9.11
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
pydantic[email]>=2.7.0
pydantic-settings>=2.2.0

# Redis for caching
redis>=5.0.3

# Utilities
python-dateutil>=2.9.0
supabase>=2.3.0
